Tests/LibWeb/Text/expected/Editing/clipboard-shortcut-paste.txt

paste event data: "PASTED"
editor text after paste: "aPASTEDb"
canceled paste event data: "PASTED"
editor text after canceled paste: "aPASTEDb"
